Move email checker within Celery tasks module
Initially the email check and notify was put in management commands
in order to test that through management commands. Now since we
have celery infrastructure ready, the task has been moved to
tasks.py
